Update to AgateLib 2.0.161.

+++ RigidBodyDynamics.Demo/ChainOnCircleExample.cs
@@ -2,7 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using AgateLib.DisplayLib;
-using AgateLib.Geometry;
+using AgateLib.Mathematics;
+using AgateLib.Mathematics.Geometry;
 
 namespace RigidBodyDynamics.Demo
 {
@@ -25,7 +26,7 @@ namespace RigidBodyDynamics.Demo
 
 		public string Name => "Chain on a circle";
 
-		public float PotentialEnergy => system.Particles.Sum(p => p.Mass * p.Position.Y * -gravity);
+		public double PotentialEnergy => system.Particles.Sum(p => p.Mass * p.Position.Y * -gravity);
 
 		public int BoxCount { get; set; } = 2;
 
@@ -66,10 +67,10 @@ namespace RigidBodyDynamics.Demo
 
 				Display.DrawEllipse(
 					(Rectangle) RectangleF.FromLTRB(
-						circlePosition.X - circleRadius,
-						circlePosition.Y - circleRadius,
-						circlePosition.X + circleRadius,
-						circlePosition.Y + circleRadius), Color.Blue);
+						(float)circlePosition.X - circleRadius,
+						(float)circlePosition.Y - circleRadius,
+						(float)circlePosition.X + circleRadius,
+						(float)circlePosition.Y + circleRadius), Color.Blue);
 			}
 		}
