+++ RigidBodyDynamics.Demo/ParticleOnCircleOffCenterExample.cs
@@ -2,7 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using AgateLib.DisplayLib;
-using AgateLib.Geometry;
+using AgateLib.Mathematics;
+using AgateLib.Mathematics.Geometry;
 
 namespace RigidBodyDynamics.Demo
 {
@@ -27,7 +28,7 @@ namespace RigidBodyDynamics.Demo
 
 		public string Name => "Particle corner on a circle";
 
-		public float PotentialEnergy => system.Particles.Sum(p => p.Mass * p.Position.Y * -gravity);
+		public double PotentialEnergy => system.Particles.Sum(p => p.Mass * p.Position.Y * -gravity);
 
 		public KinematicsSystem Initialize(Size area)
 		{
@@ -70,10 +71,10 @@ namespace RigidBodyDynamics.Demo
 
 			Display.DrawEllipse(
 				(Rectangle)RectangleF.FromLTRB(
-					circlePosition.X - circleRadius,
-					circlePosition.Y - circleRadius,
-					circlePosition.X + circleRadius,
-					circlePosition.Y + circleRadius), Color.Blue);
+					(float)circlePosition.X - circleRadius,
+					(float)circlePosition.Y - circleRadius,
+					(float)circlePosition.X + circleRadius,
+					(float)circlePosition.Y + circleRadius), Color.Blue);
 		}
 
 		private void InitializeImages()
