+++ RigidBodyDynamics/Constraints/CirclePerimeterOffcenterConstraint.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using AgateLib.Geometry;
+using AgateLib.Mathematics;
 
 namespace RigidBodyDynamics
 {
@@ -35,7 +35,7 @@ namespace RigidBodyDynamics
 
 		private Vector2 OffsetDerivative => offset.Rotate(particle.Angle + (float)Math.PI * .5f);
 
-		public float Value => .5f * (ConstrainedPointPosition.MagnitudeSquared - circleRadius * circleRadius);
+		public double Value => .5f * (ConstrainedPointPosition.MagnitudeSquared - circleRadius * circleRadius);
 
 		public bool AppliesTo(PhysicalParticle particle)
 		{
