+++ RigidBodyDynamics/VectorExtensions.cs
@@ -3,20 +3,20 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using AgateLib.Geometry;
+using AgateLib.Mathematics;
 
 namespace RigidBodyDynamics
 {
 	public static class VectorExtensions
 	{
-		public static Vector2 Rotate(this Vector2 parent, float angle)
+		public static Vector2 Rotate(this Vector2 parent, double angle)
 		{
 			return new Vector2(
 				Math.Cos(angle) * parent.X + Math.Sin(angle) * parent.Y,
 				-Math.Sin(angle) * parent.X + Math.Cos(angle) * parent.Y);
 		}
 
-		public static Vector2 RotationDerivative(this Vector2 parent, float angle)
+		public static Vector2 RotationDerivative(this Vector2 parent, double angle)
 		{
 			return new Vector2(
 				-Math.Sin(angle) * parent.X + Math.Cos(angle) * parent.Y,
